refactor: moved scripts to problems package
