+++ .github/workflows/deployment.yml
@@ -79,7 +79,8 @@ jobs:
           node-version: 22.14.0
       - uses: actions/checkout@v4
       - run: npm install
-      - run: npm run problems:cache:website
+      - run: npm run cache:website
+        working-directory: ./packages/problems
       - run: npm run deploy
         working-directory: ./apps/website
         env:
