+++ apps/queue-processor/package.json
@@ -9,11 +9,12 @@
     "execa": "^9.5.2"
   },
   "devDependencies": {
+    "@easyshell/problems": "0.1.0",
     "@types/node": "^20.17.16",
     "tsx": "^4.19.3"
   },
   "scripts": {
-    "problems:cache": "cd ../../scripts && APP=script tsx ./problems-cache-queue-processor.ts",
+    "cache": "cd ../../packages/problems && npm run cache:queue-processor",
     "build": "npx esbuild index.ts --bundle --platform=node --outfile=queue-processor.cjs",
     "start": "node queue-processor.cjs"
   }
