+++ apps/website/package.json
@@ -8,12 +8,13 @@
     "preview": "npm run pages:build && wrangler pages dev --ip 0.0.0.0",
     "deploy": "npm run pages:build && wrangler pages deploy",
     "build": "SKIP_ENV_VALIDATION=1 next build",
-    "problems:cache": "cd ../../scripts && APP=script tsx ./problems-cache-website.ts",
+    "cache": "cd ../../packages/problems && npm run cache:website",
     "dev": "next dev",
     "lint": "next lint",
     "start": "next start"
   },
   "dependencies": {
+    "@easyshell/problems": "0.1.0",
     "@auth/drizzle-adapter": "^1.7.4",
     "@easyshell/db": "0.1.0",
     "@easyshell/env": "0.1.0",
