+++ packages/problems/scripts/_utils.ts
