+++ packages/problems/scripts/build.ts
@@ -3,7 +3,7 @@ import { getProblemInfo, getProblems } from "@easyshell/problems"
 import { PROBLEMS_DIR, PROJECT_ROOT } from "@easyshell/utils/build"
 
 import { RunParallelStuff, Task } from "./_utils"
-import "./problems-lint"
+import "./lint"
 
 import { $ } from "execa"
 import { writeFile } from "fs/promises"
