+++ packages/problems/scripts/lint.ts
