+++ packages/problems/scripts/new.ts
