+++ packages/problems/scripts/test.ts
