--- scripts/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "@easyshell/scripts",
-  "version": "0.1.0",
-  "private": true,
-  "type": "module",
-  "module": "nodenext",
-  "target": "es2022",
-  "dependencies": {
-    "@easyshell/problems": "0.1.0",
-    "@easyshell/queue-processor": "0.1.0",
-    "@easyshell/utils": "0.1.0",
-    "arg": "^5.0.2",
-    "execa": "^9.5.2"
-  },
-  "devDependencies": {
-    "@types/node": "^22.13.9",
-    "tsx": "^4.19.3"
-  }
-}
